[TEST] Consolidate and expand Namediff tests to 100% coverage
- Add tests for Card object initialization and b-side recursion in Namediff.
- Add mocks for tqdm ImportError in nearest_card_par.
- Merge tests/test_namediff_gaps.py into tests/test_namediff.py.
- Remove redundant test files.
